refactor: ApplicationsPanel 서버 셸과 클라이언트 파츠 분리(#378)
- ApplicationsPanel을 서버 컴포넌트로 전환해 초기 지원 목록 조회와 기간 범위 계산 책임을 담당하도록 정리
- 탭 전환, 프리뷰, 무한 스크롤, 상태 낙관 갱신, URL 동기화 로직을 ApplicationsPanelClient로 분리
- ApplicationsView의 중간 서버 섹션을 제거해 Suspense 내부 렌더링 흐름을 단순화
- 초기 조회와 추가 페이지 조회가 같은 periodStart/periodEnd 기준을 사용하도록 데이터 흐름을 명시화
